feat: breaking refactor to events for better discoverability and relaxed restrictions around event data
