+++ packages/core/src/nodes.rs
@@ -5,10 +5,11 @@
 use crate::{
     innerlude::{AttributeValue, ComponentPtr, Element, Properties, Scope, ScopeId, ScopeState},
     lazynodes::LazyNodes,
-    AnyEvent, Component,
+    Component,
 };
 use bumpalo::{boxed::Box as BumpBox, Bump};
 use std::{
+    any::Any,
     cell::{Cell, RefCell},
     fmt::{Arguments, Debug, Formatter},
 };
@@ -373,8 +374,21 @@ pub struct Listener<'bump> {
 }
 
 pub type InternalHandler<'bump> = &'bump RefCell<Option<InternalListenerCallback<'bump>>>;
-type InternalListenerCallback<'bump> = BumpBox<'bump, dyn FnMut(AnyEvent) + 'bump>;
+type InternalListenerCallback<'bump> = BumpBox<'bump, dyn FnMut(&'bump dyn Any) + 'bump>;
 type ExternalListenerCallback<'bump, T> = BumpBox<'bump, dyn FnMut(T) + 'bump>;
+
+/// A generic event type passed to Event handlers
+pub trait UiEvent: 'static + Debug + Any {
+    /// Whether or not this event can continue to bubble upwards
+    ///
+    /// By default false
+    fn bubbles(&self) -> bool {
+        false
+    }
+
+    /// Cast this event down into any
+    fn as_any(&self) -> &dyn Any;
+}
 
 /// The callback type generated by the `rsx!` macro when an `on` field is specified for components.
 ///
