+++ packages/html/src/events/animation.rs
@@ -1,0 +1,25 @@
+use super::make_listener;
+use dioxus_core::{Listener, NodeFactory};
+use std::cell::Cell;
+
+event! {
+    AnimationEvent: [
+        /// onanimationstart
+        onanimationstart
+
+        /// onanimationend
+        onanimationend
+
+        /// onanimationiteration
+        onanimationiteration
+    ];
+}
+
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+#[derive(Debug, Clone)]
+pub struct AnimationEvent {
+    pub animation_name: String,
+    pub pseudo_element: String,
+    pub elapsed_time: f32,
+    pub bubble: Cell<bool>,
+}
