+++ packages/html/src/events/selection.rs
@@ -1,0 +1,13 @@
+use super::make_listener;
+use dioxus_core::{Listener, NodeFactory};
+
+event! {
+    SelectionEvent: [
+        /// onselect
+        onselect
+    ];
+}
+
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+#[derive(Debug, Clone)]
+pub struct SelectionEvent {}
