+++ packages/html/src/events/toggle.rs
@@ -1,0 +1,11 @@
+use super::make_listener;
+use dioxus_core::{Listener, NodeFactory};
+
+event! {
+    ToggleEvent: [
+        ///
+        ontoggle
+    ];
+}
+
+pub struct ToggleEvent {}
