+++ packages/web/Cargo.toml
@@ -58,7 +58,10 @@ features = [
     "ClipboardEvent",
     "NamedNodeMap",
     "KeyboardEvent",
+    "FileList",
     "TouchEvent",
+    "DataTransfer",
+    "DragEvent",
     "WheelEvent",
     "AnimationEvent",
     "TransitionEvent",
